+++ packages/xchain-thorchain-amm/src/types.ts
@@ -1,25 +1,7 @@
 import { FeeOption } from '@xchainjs/xchain-client'
-import { CryptoAmount } from '@xchainjs/xchain-thorchain-query'
+import { CryptoAmount, LiquidityPool } from '@xchainjs/xchain-thorchain-query'
 import { Address, Asset, BaseAmount } from '@xchainjs/xchain-util'
 import { BigNumber } from 'bignumber.js'
-
-// export type PoolCache = {
-//   lastRefreshed: number
-//   pools: Record<string, LiquidityPool>
-// }
-
-// export type AsgardCache = {
-//   lastRefreshed: number
-//   inboundAddressesItems: Record<string, InboundAddressesItem>
-// }
-// export type InboundDetailCache = {
-//   lastRefreshed: number
-//   inboundDetails: Record<string, InboundDetail>
-// }
-// export type NetworkValuesCache = {
-//   lastRefreshed: number
-//   networkValues: Record<string, number>
-// }
 
 export type MidgardConfig = {
   apiRetries: number
@@ -38,14 +20,35 @@ export type ExecuteSwap = {
   waitTimeSeconds: number
 }
 
-export type SwapSubmitted = {
+export type TxSubmitted = {
   hash: string
   url: string
   waitTimeSeconds: number
 }
 
+export type AddLiquidity = {
+  asset: CryptoAmount
+  rune: CryptoAmount
+  waitTimeSeconds: number
+}
+
+export type LiquidityPosition = {
+  assetPool: LiquidityPool
+  assetAmount: CryptoAmount
+  runeAmount: CryptoAmount
+  impermanentLossProtection: number
+}
+
+export type RemoveLiquidity = {
+  asset: CryptoAmount
+  rune: CryptoAmount
+  percentage: number
+  waitTimeSeconds: number
+  asymmetricalWithdraw?: Asset
+}
+
 export type DepositParams = {
-  walletIndex?: number // send from this HD index
+  walletIndex?: number
   asset: Asset
   amount: BaseAmount
   feeOption: FeeOption
@@ -59,36 +62,6 @@ export type SwapOutput = {
 }
 
 export type UnitData = {
-  liquidityUnits: BaseAmount
-  totalUnits: BaseAmount
+  liquidityUnits: BigNumber
+  totalUnits: BigNumber
 }
-
-export type LiquidityData = {
-  // assetDeposit: BaseAmount
-  rune: BaseAmount
-  asset: BaseAmount
-}
-
-// export type Block = {
-//   current: number
-//   lastAdded: number
-//   fullProtection: number
-// }
-
-// export type Coverage = {
-//   poolRatio: BaseAmount
-// }
-
-// export type InboundDetail = {
-//   vault: string
-//   router?: string
-//   haltedChain: boolean
-//   haltedTrading: boolean
-//   haltedLP: boolean
-//   gas_rate: BigNumber
-// }
-
-// export type ChainAttributes = {
-//   blockReward: number
-//   avgBlockTimeInSecs: number
-// }