+++ packages/xchain-bitcoin/CHANGELOG.md
@@ -1,13 +1,21 @@
-# v.0.20.1 (2022-09-25)
+# v.0.20.1 (2022-09-29)
 
 ## Update
 
 - changed default haskoin url to point to ninerealms
+
+# v.0.20.1 (2022-09-29)
+
+## Update
+
+- bumped deps on xchain-utils & xchain-client
+
 # v.0.20.0 (2022-09-05)
 
 ## Update
 
 - bumped deps on xchain-utils & xchain-client
+
 # v.0.19.0 (2022-07-21)
 
 ### Breaking change