+++ packages/xchain-bitcoincash/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v.0.15.1 (2022-09-29)
+
+## Update
+
+- bumped deps on xchain-utils & xchain-client
+
 # v.0.15.0 (2022-09-05)
 
 ## Update