+++ packages/xchain-thorchain-query/CHANGELOG.md
@@ -1,3 +1,16 @@
+# v0.0.1.0-beta2 (2022-09-29)
+
+## Add
+
+- Add estimate add & withdraw Liquidity
+- Add Check Liquidity position
+
+# v0.0.1.0-beta1 (2022-09-29)
+
+## Update
+
+- bumped deps on xchain-utils & xchain-client
+
 # v0.0.1.0-beta (2022-09-15)
 
 ## Fix